rename {WindowFunctions => window_functions}.hpp
