+++ include/window_functions.hpp
